move things to utils files, add types
