+++ pages/_app.tsx
@@ -10,7 +10,7 @@ import Maintenance from 'components/pages/Error/Maintenance';
 import { getConfig } from 'global/config';
 import queryString from 'query-string';
 import { get, omit } from 'lodash';
-import { fetchEgoToken, OAUTH_QUERY_PARAM_NAME } from 'global/utils/authUtils';
+import { fetchEgoToken, makeRedirectPath, OAUTH_QUERY_PARAM_NAME } from 'global/utils/authUtils';
 import Loader from 'components/Loader';
 
 const redirectTo = (res: any, url: string) => {
@@ -22,11 +22,6 @@ const redirectTo = (res: any, url: string) => {
   } else {
     Router.push(url);
   }
-};
-
-const makeRedirectPath = (ctxAsPath: string | undefined): string => {
-  const path = ctxAsPath ? `/?redirect=${encodeURI(ctxAsPath)}` : '/';
-  return path;
 };
 
 const enforceLogin = ({ ctx }: { ctx: NextPageContext }) => {
